refactor(events.py): move DB queries from route handlers to services
Four collection/review endpoints in api/events.py were issuing raw
db.query(Event) plus owner/admin checks inline — a classic route-layer
business-logic leak. The fix is two-layered:

1. New dependency `get_event_for_dj_or_admin(code, user, db)` in
   api/deps.py. Returns 404 if missing, 403 if the user is neither the
   owner nor an admin. Handlers now take `event: Event = Depends(...)`
   and don't touch the query layer.

2. Collection logic moved to services/collect.py:
   - collection_settings_payload(event)       — response serialization
   - update_collection_settings(db, event, payload)  — mutation + validation + commit
   - get_pending_review_rows(db, event_id, limit=200)
   - execute_bulk_review(db, event_id, payload) -> (accepted, rejected)

Handlers shrank from ~170 LOC to ~45 LOC. The _to_naive_utc helper
(previously duplicated) moved to services/collect.py where it's scoped
to the collection mutations that need it.

Test suite: 128 event/collect tests green.
