+++ server/app/services/collect.py
@@ -1,9 +1,136 @@
 """Service layer for pre-event collection."""
 
+from datetime import datetime, timedelta
+
+from fastapi import HTTPException
 from sqlalchemy.orm import Session
 
 from app.models.event import Event
 from app.models.guest_profile import GuestProfile
+from app.models.request import Request as SongRequest
+from app.schemas.collect import BulkReviewRequest, UpdateCollectionSettings
+
+
+def _to_naive_utc(dt: datetime) -> datetime:
+    """Normalize a datetime to naive UTC (tzinfo stripped)."""
+    if dt.tzinfo is not None:
+        dt = dt.astimezone(tz=None).replace(tzinfo=None)
+    return dt
+
+
+def collection_settings_payload(event: Event) -> dict:
+    """Serialize collection-scheduling settings for the dashboard."""
+    return {
+        "collection_opens_at": event.collection_opens_at,
+        "live_starts_at": event.live_starts_at,
+        "submission_cap_per_guest": event.submission_cap_per_guest,
+        "collection_phase_override": event.collection_phase_override,
+        "phase": event.phase,
+    }
+
+
+def update_collection_settings(
+    db: Session, event: Event, payload: UpdateCollectionSettings
+) -> Event:
+    """Apply a collection-settings update and commit.
+
+    Validates that collection_opens_at precedes live_starts_at and auto-extends
+    expires_at when the scheduled live phase runs past it (the default 6h event
+    expiry is too short for multi-day pre-event collection).
+    """
+    if payload.collection_opens_at is not None:
+        event.collection_opens_at = _to_naive_utc(payload.collection_opens_at)
+    if payload.live_starts_at is not None:
+        event.live_starts_at = _to_naive_utc(payload.live_starts_at)
+    if payload.submission_cap_per_guest is not None:
+        event.submission_cap_per_guest = payload.submission_cap_per_guest
+    if "collection_phase_override" in payload.model_fields_set:
+        event.collection_phase_override = payload.collection_phase_override
+
+    opens = event.collection_opens_at
+    live = event.live_starts_at
+    expires = event.expires_at
+    if opens and live and opens >= live:
+        raise HTTPException(
+            status_code=400, detail="collection_opens_at must be before live_starts_at"
+        )
+    if live and expires and live >= expires:
+        event.expires_at = live + timedelta(hours=12)
+
+    db.commit()
+    db.refresh(event)
+    return event
+
+
+def get_pending_review_rows(db: Session, event_id: int, limit: int = 200) -> list[SongRequest]:
+    """Collection-phase requests awaiting DJ review, ranked by votes then age."""
+    return (
+        db.query(SongRequest)
+        .filter(SongRequest.event_id == event_id)
+        .filter(SongRequest.submitted_during_collection == True)  # noqa: E712
+        .filter(SongRequest.status == "new")
+        .order_by(SongRequest.vote_count.desc(), SongRequest.created_at.asc())
+        .limit(limit)
+        .all()
+    )
+
+
+def execute_bulk_review(db: Session, event_id: int, payload: BulkReviewRequest) -> tuple[int, int]:
+    """Apply a bulk-review action to collection-phase pending requests.
+
+    Returns (accepted_count, rejected_count). Raises HTTPException(400) when the
+    payload parameters are inconsistent with the selected action.
+    """
+    pending_q = (
+        db.query(SongRequest)
+        .filter(SongRequest.event_id == event_id)
+        .filter(SongRequest.submitted_during_collection == True)  # noqa: E712
+        .filter(SongRequest.status == "new")
+    )
+
+    accepted = 0
+    rejected = 0
+
+    if payload.action == "accept_top_n":
+        if payload.n is None:
+            raise HTTPException(status_code=400, detail="n is required")
+        rows = (
+            pending_q.order_by(SongRequest.vote_count.desc(), SongRequest.created_at.asc())
+            .limit(payload.n)
+            .all()
+        )
+        for r in rows:
+            r.status = "accepted"
+            accepted += 1
+    elif payload.action == "accept_threshold":
+        if payload.min_votes is None:
+            raise HTTPException(status_code=400, detail="min_votes is required")
+        rows = pending_q.filter(SongRequest.vote_count >= payload.min_votes).all()
+        for r in rows:
+            r.status = "accepted"
+            accepted += 1
+    elif payload.action == "accept_ids":
+        if not payload.request_ids:
+            raise HTTPException(status_code=400, detail="request_ids is required")
+        rows = pending_q.filter(SongRequest.id.in_(payload.request_ids)).all()
+        for r in rows:
+            r.status = "accepted"
+            accepted += 1
+    elif payload.action == "reject_ids":
+        if not payload.request_ids:
+            raise HTTPException(status_code=400, detail="request_ids is required")
+        rows = pending_q.filter(SongRequest.id.in_(payload.request_ids)).all()
+        for r in rows:
+            r.status = "rejected"
+            rejected += 1
+    elif payload.action == "reject_remaining":
+        rows = pending_q.all()
+        for r in rows:
+            r.status = "rejected"
+            rejected += 1
+
+    db.commit()
+    return accepted, rejected
 
 
 class SubmissionCapExceeded(Exception):
